requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
mysqlclient
redis
django-redis
django-environ